refactor(experimental): graphql: refactor resolvers into their own directory
