+++ packages/rpc-graphql/src/resolvers/account.ts
@@ -13,3 +13,88 @@ export const resolveAccount = (fieldName: string) => {
     ) =>
         parent[fieldName] === null ? null : context.loaders.account.load({ ...args, address: parent[fieldName] }, info);
 };
+
+export const accountResolvers = {
+    Account: {
+        __resolveType(account: { encoding: string; programName: string; accountType: string }) {
+            if (account.encoding === 'base58') {
+                return 'AccountBase58';
+            }
+            if (account.encoding === 'base64') {
+                return 'AccountBase64';
+            }
+            if (account.encoding === 'base64+zstd') {
+                return 'AccountBase64Zstd';
+            }
+            if (account.encoding === 'jsonParsed') {
+                if (account.programName === 'nonce') {
+                    return 'NonceAccount';
+                }
+                if (account.accountType === 'mint' && account.programName === 'spl-token') {
+                    return 'MintAccount';
+                }
+                if (account.accountType === 'account' && account.programName === 'spl-token') {
+                    return 'TokenAccount';
+                }
+                if (account.programName === 'stake') {
+                    return 'StakeAccount';
+                }
+                if (account.accountType === 'vote' && account.programName === 'vote') {
+                    return 'VoteAccount';
+                }
+                if (account.accountType === 'lookupTable' && account.programName === 'address-lookup-table') {
+                    return 'LookupTableAccount';
+                }
+            }
+            return 'AccountBase64';
+        },
+    },
+    AccountBase58: {
+        ownerProgram: resolveAccount('ownerProgram'),
+    },
+    AccountBase64: {
+        ownerProgram: resolveAccount('ownerProgram'),
+    },
+    AccountBase64Zstd: {
+        ownerProgram: resolveAccount('ownerProgram'),
+    },
+    LookupTableAccount: {
+        authority: resolveAccount('authority'),
+        ownerProgram: resolveAccount('ownerProgram'),
+    },
+    MintAccount: {
+        freezeAuthority: resolveAccount('freezeAuthority'),
+        mintAuthority: resolveAccount('mintAuthority'),
+        ownerProgram: resolveAccount('ownerProgram'),
+    },
+    NonceAccount: {
+        authority: resolveAccount('authority'),
+        ownerProgram: resolveAccount('ownerProgram'),
+    },
+    StakeAccount: {
+        ownerProgram: resolveAccount('ownerProgram'),
+    },
+    StakeAccountDataMetaAuthorized: {
+        staker: resolveAccount('staker'),
+        withdrawer: resolveAccount('withdrawer'),
+    },
+    StakeAccountDataMetaLockup: {
+        custodian: resolveAccount('custodian'),
+    },
+    StakeAccountDataStakeDelegation: {
+        voter: resolveAccount('voter'),
+    },
+    TokenAccount: {
+        mint: resolveAccount('mint'),
+        owner: resolveAccount('owner'),
+        ownerProgram: resolveAccount('ownerProgram'),
+    },
+    VoteAccount: {
+        authorizedWithdrawer: resolveAccount('authorizedWithdrawer'),
+        node: resolveAccount('nodePubkey'),
+        ownerProgram: resolveAccount('ownerProgram'),
+    },
+    VoteAccountDataAuthorizedVoter: {
+        authorizedVoter: resolveAccount('authorizedVoter'),
+    },
+};
