+++ packages/rpc-graphql/src/schema/account.ts
@@ -1,6 +1,3 @@
-/* eslint-disable sort-keys-fix/sort-keys-fix */
-import { resolveAccount } from '../resolvers/account';
-
 export const accountTypeDefs = /* GraphQL */ `
     """
     Account interface
@@ -199,88 +196,3 @@ export const accountTypeDefs = /* GraphQL */ `
         votes: [VoteAccountDataVote]
     }
 `;
-
-export const accountResolvers = {
-    Account: {
-        __resolveType(account: { encoding: string; programName: string; accountType: string }) {
-            if (account.encoding === 'base58') {
-                return 'AccountBase58';
-            }
-            if (account.encoding === 'base64') {
-                return 'AccountBase64';
-            }
-            if (account.encoding === 'base64+zstd') {
-                return 'AccountBase64Zstd';
-            }
-            if (account.encoding === 'jsonParsed') {
-                if (account.programName === 'nonce') {
-                    return 'NonceAccount';
-                }
-                if (account.accountType === 'mint' && account.programName === 'spl-token') {
-                    return 'MintAccount';
-                }
-                if (account.accountType === 'account' && account.programName === 'spl-token') {
-                    return 'TokenAccount';
-                }
-                if (account.programName === 'stake') {
-                    return 'StakeAccount';
-                }
-                if (account.accountType === 'vote' && account.programName === 'vote') {
-                    return 'VoteAccount';
-                }
-                if (account.accountType === 'lookupTable' && account.programName === 'address-lookup-table') {
-                    return 'LookupTableAccount';
-                }
-            }
-            return 'AccountBase64';
-        },
-    },
-    AccountBase58: {
-        ownerProgram: resolveAccount('ownerProgram'),
-    },
-    AccountBase64: {
-        ownerProgram: resolveAccount('ownerProgram'),
-    },
-    AccountBase64Zstd: {
-        ownerProgram: resolveAccount('ownerProgram'),
-    },
-    NonceAccount: {
-        authority: resolveAccount('authority'),
-        ownerProgram: resolveAccount('ownerProgram'),
-    },
-    LookupTableAccount: {
-        authority: resolveAccount('authority'),
-        ownerProgram: resolveAccount('ownerProgram'),
-    },
-    MintAccount: {
-        freezeAuthority: resolveAccount('freezeAuthority'),
-        mintAuthority: resolveAccount('mintAuthority'),
-        ownerProgram: resolveAccount('ownerProgram'),
-    },
-    TokenAccount: {
-        mint: resolveAccount('mint'),
-        owner: resolveAccount('owner'),
-        ownerProgram: resolveAccount('ownerProgram'),
-    },
-    StakeAccountDataMetaAuthorized: {
-        staker: resolveAccount('staker'),
-        withdrawer: resolveAccount('withdrawer'),
-    },
-    StakeAccountDataMetaLockup: {
-        custodian: resolveAccount('custodian'),
-    },
-    StakeAccountDataStakeDelegation: {
-        voter: resolveAccount('voter'),
-    },
-    StakeAccount: {
-        ownerProgram: resolveAccount('ownerProgram'),
-    },
-    VoteAccountDataAuthorizedVoter: {
-        authorizedVoter: resolveAccount('authorizedVoter'),
-    },
-    VoteAccount: {
-        authorizedWithdrawer: resolveAccount('authorizedWithdrawer'),
-        node: resolveAccount('nodePubkey'),
-        ownerProgram: resolveAccount('ownerProgram'),
-    },
-};
