+++ packages/rpc-graphql/src/schema/instruction.ts
@@ -1,11 +1,6 @@
-/* eslint-disable sort-keys-fix/sort-keys-fix */
-
 // *Note: Any integer value inside a transaction instruction must be `BigInt`
 // until the following issue is resolved:
 // <https://github.com/solana-labs/solana-web3.js/issues/1828>
-
-import { resolveAccount } from '../resolvers/account';
-
 export const instructionTypeDefs = /* GraphQL */ `
     """
     Transaction instruction interface
@@ -1049,719 +1044,3 @@ export const instructionTypeDefs = /* GraphQL */ `
         voteAccount: Account
     }
 `;
-
-export const instructionResolvers = {
-    TransactionInstruction: {
-        __resolveType(instruction: { programName: string; instructionType: string }) {
-            if (instruction.programName) {
-                if (instruction.programName === 'address-lookup-table') {
-                    if (instruction.instructionType === 'createLookupTable') {
-                        return 'CreateLookupTableInstruction';
-                    }
-                    if (instruction.instructionType === 'freezeLookupTable') {
-                        return 'FreezeLookupTableInstruction';
-                    }
-                    if (instruction.instructionType === 'extendLookupTable') {
-                        return 'ExtendLookupTableInstruction';
-                    }
-                    if (instruction.instructionType === 'deactivateLookupTable') {
-                        return 'DeactivateLookupTableInstruction';
-                    }
-                    if (instruction.instructionType === 'closeLookupTable') {
-                        return 'CloseLookupTableInstruction';
-                    }
-                }
-                if (instruction.programName === 'bpf-loader') {
-                    if (instruction.instructionType === 'write') {
-                        return 'BpfLoaderWriteInstruction';
-                    }
-                    if (instruction.instructionType === 'finalize') {
-                        return 'BpfLoaderFinalizeInstruction';
-                    }
-                }
-                if (instruction.programName === 'bpf-upgradeable-loader') {
-                    if (instruction.instructionType === 'initializeBuffer') {
-                        return 'BpfUpgradeableLoaderInitializeBufferInstruction';
-                    }
-                    if (instruction.instructionType === 'write') {
-                        return 'BpfUpgradeableLoaderWriteInstruction';
-                    }
-                    if (instruction.instructionType === 'deployWithMaxDataLen') {
-                        return 'BpfUpgradeableLoaderDeployWithMaxDataLenInstruction';
-                    }
-                    if (instruction.instructionType === 'upgrade') {
-                        return 'BpfUpgradeableLoaderUpgradeInstruction';
-                    }
-                    if (instruction.instructionType === 'setAuthority') {
-                        return 'BpfUpgradeableLoaderSetAuthorityInstruction';
-                    }
-                    if (instruction.instructionType === 'setAuthorityChecked') {
-                        return 'BpfUpgradeableLoaderSetAuthorityCheckedInstruction';
-                    }
-                    if (instruction.instructionType === 'close') {
-                        return 'BpfUpgradeableLoaderCloseInstruction';
-                    }
-                    if (instruction.instructionType === 'extendProgram') {
-                        return 'BpfUpgradeableLoaderExtendProgramInstruction';
-                    }
-                }
-                if (instruction.programName === 'spl-associated-token-account') {
-                    if (instruction.instructionType === 'create') {
-                        return 'SplAssociatedTokenCreateInstruction';
-                    }
-                    if (instruction.instructionType === 'createIdempotent') {
-                        return 'SplAssociatedTokenCreateIdempotentInstruction';
-                    }
-                    if (instruction.instructionType === 'recoverNested') {
-                        return 'SplAssociatedTokenRecoverNestedInstruction';
-                    }
-                }
-                if (instruction.programName === 'spl-memo') {
-                    return 'SplMemoInstruction';
-                }
-                if (instruction.programName === 'spl-token') {
-                    if (instruction.instructionType === 'initializeMint') {
-                        return 'SplTokenInitializeMintInstruction';
-                    }
-                    if (instruction.instructionType === 'initializeMint2') {
-                        return 'SplTokenInitializeMint2Instruction';
-                    }
-                    if (instruction.instructionType === 'initializeAccount') {
-                        return 'SplTokenInitializeAccountInstruction';
-                    }
-                    if (instruction.instructionType === 'initializeAccount2') {
-                        return 'SplTokenInitializeAccount2Instruction';
-                    }
-                    if (instruction.instructionType === 'initializeAccount3') {
-                        return 'SplTokenInitializeAccount3Instruction';
-                    }
-                    if (instruction.instructionType === 'initializeMultisig') {
-                        return 'SplTokenInitializeMultisigInstruction';
-                    }
-                    if (instruction.instructionType === 'initializeMultisig2') {
-                        return 'SplTokenInitializeMultisig2Instruction';
-                    }
-                    if (instruction.instructionType === 'transfer') {
-                        return 'SplTokenTransferInstruction';
-                    }
-                    if (instruction.instructionType === 'approve') {
-                        return 'SplTokenApproveInstruction';
-                    }
-                    if (instruction.instructionType === 'revoke') {
-                        return 'SplTokenRevokeInstruction';
-                    }
-                    if (instruction.instructionType === 'setAuthority') {
-                        return 'SplTokenSetAuthorityInstruction';
-                    }
-                    if (instruction.instructionType === 'mintTo') {
-                        return 'SplTokenMintToInstruction';
-                    }
-                    if (instruction.instructionType === 'burn') {
-                        return 'SplTokenBurnInstruction';
-                    }
-                    if (instruction.instructionType === 'closeAccount') {
-                        return 'SplTokenCloseAccountInstruction';
-                    }
-                    if (instruction.instructionType === 'freezeAccount') {
-                        return 'SplTokenFreezeAccountInstruction';
-                    }
-                    if (instruction.instructionType === 'thawAccount') {
-                        return 'SplTokenThawAccountInstruction';
-                    }
-                    if (instruction.instructionType === 'transferChecked') {
-                        return 'SplTokenTransferCheckedInstruction';
-                    }
-                    if (instruction.instructionType === 'approveChecked') {
-                        return 'SplTokenApproveCheckedInstruction';
-                    }
-                    if (instruction.instructionType === 'mintToChecked') {
-                        return 'SplTokenMintToCheckedInstruction';
-                    }
-                    if (instruction.instructionType === 'burnChecked') {
-                        return 'SplTokenBurnCheckedInstruction';
-                    }
-                    if (instruction.instructionType === 'syncNative') {
-                        return 'SplTokenSyncNativeInstruction';
-                    }
-                    if (instruction.instructionType === 'getAccountDataSize') {
-                        return 'SplTokenGetAccountDataSizeInstruction';
-                    }
-                    if (instruction.instructionType === 'initializeImmutableOwner') {
-                        return 'SplTokenInitializeImmutableOwnerInstruction';
-                    }
-                    if (instruction.instructionType === 'amountToUiAmount') {
-                        return 'SplTokenAmountToUiAmountInstruction';
-                    }
-                    if (instruction.instructionType === 'uiAmountToAmount') {
-                        return 'SplTokenUiAmountToAmountInstruction';
-                    }
-                    if (instruction.instructionType === 'initializeMintCloseAuthority') {
-                        return 'SplTokenInitializeMintCloseAuthorityInstruction';
-                    }
-                }
-                if (instruction.programName === 'stake') {
-                    if (instruction.instructionType === 'initialize') {
-                        return 'StakeInitializeInstruction';
-                    }
-                    if (instruction.instructionType === 'authorize') {
-                        return 'StakeAuthorizeInstruction';
-                    }
-                    if (instruction.instructionType === 'delegate') {
-                        return 'StakeDelegateStakeInstruction';
-                    }
-                    if (instruction.instructionType === 'split') {
-                        return 'StakeSplitInstruction';
-                    }
-                    if (instruction.instructionType === 'withdraw') {
-                        return 'StakeWithdrawInstruction';
-                    }
-                    if (instruction.instructionType === 'deactivate') {
-                        return 'StakeDeactivateInstruction';
-                    }
-                    if (instruction.instructionType === 'setLockup') {
-                        return 'StakeSetLockupInstruction';
-                    }
-                    if (instruction.instructionType === 'merge') {
-                        return 'StakeMergeInstruction';
-                    }
-                    if (instruction.instructionType === 'authorizeWithSeed') {
-                        return 'StakeAuthorizeWithSeedInstruction';
-                    }
-                    if (instruction.instructionType === 'initializeChecked') {
-                        return 'StakeInitializeCheckedInstruction';
-                    }
-                    if (instruction.instructionType === 'authorizeChecked') {
-                        return 'StakeAuthorizeCheckedInstruction';
-                    }
-                    if (instruction.instructionType === 'authorizeCheckedWithSeed') {
-                        return 'StakeAuthorizeCheckedWithSeedInstruction';
-                    }
-                    if (instruction.instructionType === 'setLockupChecked') {
-                        return 'StakeSetLockupCheckedInstruction';
-                    }
-                    if (instruction.instructionType === 'deactivateDelinquent') {
-                        return 'StakeDeactivateDelinquentInstruction';
-                    }
-                    if (instruction.instructionType === 'redelegate') {
-                        return 'StakeRedelegateInstruction';
-                    }
-                }
-                if (instruction.programName === 'system') {
-                    if (instruction.instructionType === 'createAccount') {
-                        return 'CreateAccountInstruction';
-                    }
-                    if (instruction.instructionType === 'assign') {
-                        return 'AssignInstruction';
-                    }
-                    if (instruction.instructionType === 'transfer') {
-                        return 'TransferInstruction';
-                    }
-                    if (instruction.instructionType === 'createAccountWithSeed') {
-                        return 'CreateAccountWithSeedInstruction';
-                    }
-                    if (instruction.instructionType === 'advanceNonceAccount') {
-                        return 'AdvanceNonceAccountInstruction';
-                    }
-                    if (instruction.instructionType === 'withdrawNonceAccount') {
-                        return 'WithdrawNonceAccountInstruction';
-                    }
-                    if (instruction.instructionType === 'initializeNonceAccount') {
-                        return 'InitializeNonceAccountInstruction';
-                    }
-                    if (instruction.instructionType === 'authorizeNonceAccount') {
-                        return 'AuthorizeNonceAccountInstruction';
-                    }
-                    if (instruction.instructionType === 'upgradeNonceAccount') {
-                        return 'UpgradeNonceAccountInstruction';
-                    }
-                    if (instruction.instructionType === 'allocate') {
-                        return 'AllocateInstruction';
-                    }
-                    if (instruction.instructionType === 'allocateWithSeed') {
-                        return 'AllocateWithSeedInstruction';
-                    }
-                    if (instruction.instructionType === 'assignWithSeed') {
-                        return 'AssignWithSeedInstruction';
-                    }
-                    if (instruction.instructionType === 'transferWithSeed') {
-                        return 'TransferWithSeedInstruction';
-                    }
-                }
-                if (instruction.programName === 'vote') {
-                    if (instruction.instructionType === 'initialize') {
-                        return 'VoteInitializeAccountInstruction';
-                    }
-                    if (instruction.instructionType === 'authorize') {
-                        return 'VoteAuthorizeInstruction';
-                    }
-                    if (instruction.instructionType === 'authorizeWithSeed') {
-                        return 'VoteAuthorizeWithSeedInstruction';
-                    }
-                    if (instruction.instructionType === 'authorizeCheckedWithSeed') {
-                        return 'VoteAuthorizeCheckedWithSeedInstruction';
-                    }
-                    if (instruction.instructionType === 'vote') {
-                        return 'VoteVoteInstruction';
-                    }
-                    if (instruction.instructionType === 'updatevotestate') {
-                        return 'VoteUpdateVoteStateInstruction';
-                    }
-                    if (instruction.instructionType === 'updatevotestateswitch') {
-                        return 'VoteUpdateVoteStateSwitchInstruction';
-                    }
-                    if (instruction.instructionType === 'compactupdatevotestate') {
-                        return 'VoteCompactUpdateVoteStateInstruction';
-                    }
-                    if (instruction.instructionType === 'compactupdatevotestateswitch') {
-                        return 'VoteCompactUpdateVoteStateSwitchInstruction';
-                    }
-                    if (instruction.instructionType === 'withdraw') {
-                        return 'VoteWithdrawInstruction';
-                    }
-                    if (instruction.instructionType === 'updateValidatorIdentity') {
-                        return 'VoteUpdateValidatorIdentityInstruction';
-                    }
-                    if (instruction.instructionType === 'updateCommission') {
-                        return 'VoteUpdateCommissionInstruction';
-                    }
-                    if (instruction.instructionType === 'voteSwitch') {
-                        return 'VoteVoteSwitchInstruction';
-                    }
-                    if (instruction.instructionType === 'authorizeChecked') {
-                        return 'VoteAuthorizeCheckedInstruction';
-                    }
-                }
-            }
-            return 'GenericInstruction';
-        },
-    },
-    CreateLookupTableInstruction: {
-        lookupTableAccount: resolveAccount('lookupTableAccount'),
-        lookupTableAuthority: resolveAccount('lookupTableAuthority'),
-        payerAccount: resolveAccount('payerAccount'),
-        systemProgram: resolveAccount('systemProgram'),
-    },
-    ExtendLookupTableInstruction: {
-        lookupTableAccount: resolveAccount('lookupTableAccount'),
-        lookupTableAuthority: resolveAccount('lookupTableAuthority'),
-        payerAccount: resolveAccount('payerAccount'),
-        systemProgram: resolveAccount('systemProgram'),
-    },
-    FreezeLookupTableInstruction: {
-        lookupTableAccount: resolveAccount('lookupTableAccount'),
-        lookupTableAuthority: resolveAccount('lookupTableAuthority'),
-    },
-    DeactivateLookupTableInstruction: {
-        lookupTableAccount: resolveAccount('lookupTableAccount'),
-        lookupTableAuthority: resolveAccount('lookupTableAuthority'),
-    },
-    CloseLookupTableInstruction: {
-        lookupTableAccount: resolveAccount('lookupTableAccount'),
-        lookupTableAuthority: resolveAccount('lookupTableAuthority'),
-        recipient: resolveAccount('recipient'),
-    },
-    BpfLoaderWriteInstruction: {
-        account: resolveAccount('account'),
-    },
-    BpfLoaderFinalizeInstruction: {
-        account: resolveAccount('account'),
-    },
-    BpfUpgradeableLoaderInitializeBufferInstruction: {
-        account: resolveAccount('account'),
-    },
-    BpfUpgradeableLoaderWriteInstruction: {
-        account: resolveAccount('account'),
-        authority: resolveAccount('authority'),
-    },
-    BpfUpgradeableLoaderDeployWithMaxDataLenInstruction: {
-        authority: resolveAccount('authority'),
-        bufferAccount: resolveAccount('bufferAccount'),
-        payerAccount: resolveAccount('payerAccount'),
-        programAccount: resolveAccount('programAccount'),
-        programDataAccount: resolveAccount('programDataAccount'),
-    },
-    BpfUpgradeableLoaderUpgradeInstruction: {
-        authority: resolveAccount('authority'),
-        bufferAccount: resolveAccount('bufferAccount'),
-        programAccount: resolveAccount('programAccount'),
-        programDataAccount: resolveAccount('programDataAccount'),
-    },
-    BpfUpgradeableLoaderSetAuthorityInstruction: {
-        account: resolveAccount('account'),
-        authority: resolveAccount('authority'),
-        newAuthority: resolveAccount('newAuthority'),
-    },
-    BpfUpgradeableLoaderSetAuthorityCheckedInstruction: {
-        account: resolveAccount('account'),
-        authority: resolveAccount('authority'),
-        newAuthority: resolveAccount('newAuthority'),
-    },
-    BpfUpgradeableLoaderCloseInstruction: {
-        account: resolveAccount('account'),
-        authority: resolveAccount('authority'),
-        programAccount: resolveAccount('programAccount'),
-        recipient: resolveAccount('recipient'),
-    },
-    BpfUpgradeableLoaderExtendProgramInstruction: {
-        payerAccount: resolveAccount('payerAccount'),
-        programAccount: resolveAccount('programAccount'),
-        programDataAccount: resolveAccount('programDataAccount'),
-        systemProgram: resolveAccount('systemProgram'),
-    },
-    SplAssociatedTokenCreateInstruction: {
-        account: resolveAccount('account'),
-        mint: resolveAccount('mint'),
-        source: resolveAccount('source'),
-        systemProgram: resolveAccount('systemProgram'),
-        tokenProgram: resolveAccount('tokenProgram'),
-        wallet: resolveAccount('wallet'),
-    },
-    SplAssociatedTokenCreateIdempotentInstruction: {
-        account: resolveAccount('account'),
-        mint: resolveAccount('mint'),
-        source: resolveAccount('source'),
-        systemProgram: resolveAccount('systemProgram'),
-        tokenProgram: resolveAccount('tokenProgram'),
-        wallet: resolveAccount('wallet'),
-    },
-    SplAssociatedTokenRecoverNestedInstruction: {
-        destination: resolveAccount('destination'),
-        nestedMint: resolveAccount('nestedMint'),
-        nestedOwner: resolveAccount('nestedOwner'),
-        nestedSource: resolveAccount('nestedSource'),
-        ownerMint: resolveAccount('ownerMint'),
-        tokenProgram: resolveAccount('tokenProgram'),
-        wallet: resolveAccount('wallet'),
-    },
-    SplTokenInitializeMintInstruction: {
-        freezeAuthority: resolveAccount('freezeAuthority'),
-        mint: resolveAccount('mint'),
-        mintAuthority: resolveAccount('mintAuthority'),
-    },
-    SplTokenInitializeMint2Instruction: {
-        freezeAuthority: resolveAccount('freezeAuthority'),
-        mint: resolveAccount('mint'),
-        mintAuthority: resolveAccount('mintAuthority'),
-    },
-    SplTokenInitializeAccountInstruction: {
-        account: resolveAccount('account'),
-        mint: resolveAccount('mint'),
-        owner: resolveAccount('owner'),
-    },
-    SplTokenInitializeAccount2Instruction: {
-        account: resolveAccount('account'),
-        mint: resolveAccount('mint'),
-        owner: resolveAccount('owner'),
-    },
-    SplTokenInitializeAccount3Instruction: {
-        account: resolveAccount('account'),
-        mint: resolveAccount('mint'),
-        owner: resolveAccount('owner'),
-    },
-    SplTokenInitializeMultisigInstruction: {
-        multisig: resolveAccount('multisig'),
-    },
-    SplTokenInitializeMultisig2Instruction: {
-        multisig: resolveAccount('multisig'),
-    },
-    SplTokenTransferInstruction: {
-        authority: resolveAccount('authority'),
-        destination: resolveAccount('destination'),
-        multisigAuthority: resolveAccount('multisigAuthority'),
-        source: resolveAccount('source'),
-    },
-    SplTokenApproveInstruction: {
-        delegate: resolveAccount('delegate'),
-        multisigOwner: resolveAccount('multisigOwner'),
-        owner: resolveAccount('owner'),
-        source: resolveAccount('source'),
-    },
-    SplTokenRevokeInstruction: {
-        multisigOwner: resolveAccount('multisigOwner'),
-        owner: resolveAccount('owner'),
-        source: resolveAccount('source'),
-    },
-    SplTokenSetAuthorityInstruction: {
-        authority: resolveAccount('authority'),
-        multisigAuthority: resolveAccount('multisigAuthority'),
-        newAuthority: resolveAccount('newAuthority'),
-    },
-    SplTokenMintToInstruction: {
-        account: resolveAccount('account'),
-        authority: resolveAccount('authority'),
-        mint: resolveAccount('mint'),
-        mintAuthority: resolveAccount('mintAuthority'),
-        multisigMintAuthority: resolveAccount('multisigMintAuthority'),
-    },
-    SplTokenBurnInstruction: {
-        account: resolveAccount('account'),
-        authority: resolveAccount('authority'),
-        mint: resolveAccount('mint'),
-        multisigAuthority: resolveAccount('multisigAuthority'),
-    },
-    SplTokenCloseAccountInstruction: {
-        account: resolveAccount('account'),
-        destination: resolveAccount('destination'),
-        multisigOwner: resolveAccount('multisigOwner'),
-        owner: resolveAccount('owner'),
-    },
-    SplTokenFreezeAccountInstruction: {
-        account: resolveAccount('account'),
-        freezeAuthority: resolveAccount('freezeAuthority'),
-        mint: resolveAccount('mint'),
-        multisigFreezeAuthority: resolveAccount('multisigFreezeAuthority'),
-    },
-    SplTokenThawAccountInstruction: {
-        account: resolveAccount('account'),
-        freezeAuthority: resolveAccount('freezeAuthority'),
-        mint: resolveAccount('mint'),
-        multisigFreezeAuthority: resolveAccount('multisigFreezeAuthority'),
-    },
-    SplTokenTransferCheckedInstruction: {
-        authority: resolveAccount('authority'),
-        destination: resolveAccount('destination'),
-        mint: resolveAccount('mint'),
-        multisigAuthority: resolveAccount('multisigAuthority'),
-        source: resolveAccount('source'),
-    },
-    SplTokenApproveCheckedInstruction: {
-        delegate: resolveAccount('delegate'),
-        mint: resolveAccount('mint'),
-        multisigOwner: resolveAccount('multisigOwner'),
-        owner: resolveAccount('owner'),
-        source: resolveAccount('source'),
-    },
-    SplTokenMintToCheckedInstruction: {
-        account: resolveAccount('account'),
-        authority: resolveAccount('authority'),
-        mint: resolveAccount('mint'),
-        mintAuthority: resolveAccount('mintAuthority'),
-        multisigMintAuthority: resolveAccount('multisigMintAuthority'),
-    },
-    SplTokenBurnCheckedInstruction: {
-        account: resolveAccount('account'),
-        authority: resolveAccount('authority'),
-        mint: resolveAccount('mint'),
-        multisigAuthority: resolveAccount('multisigAuthority'),
-    },
-    SplTokenSyncNativeInstruction: {
-        account: resolveAccount('account'),
-    },
-    SplTokenGetAccountDataSizeInstruction: {
-        mint: resolveAccount('mint'),
-    },
-    SplTokenAmountToUiAmountInstruction: {
-        mint: resolveAccount('mint'),
-    },
-    SplTokenUiAmountToAmountInstruction: {
-        mint: resolveAccount('mint'),
-    },
-    SplTokenInitializeMintCloseAuthorityInstruction: {
-        mint: resolveAccount('mint'),
-        newAuthority: resolveAccount('newAuthority'),
-    },
-    Lockup: {
-        custodian: resolveAccount('custodian'),
-    },
-    StakeInitializeInstructionDataAuthorized: {
-        staker: resolveAccount('staker'),
-        withdrawer: resolveAccount('withdrawer'),
-    },
-    StakeInitializeInstruction: {
-        stakeAccount: resolveAccount('stakeAccount'),
-    },
-    StakeAuthorizeInstruction: {
-        authority: resolveAccount('authority'),
-        custodian: resolveAccount('custodian'),
-        newAuthority: resolveAccount('newAuthority'),
-        stakeAccount: resolveAccount('stakeAccount'),
-    },
-    StakeDelegateStakeInstruction: {
-        stakeAccount: resolveAccount('stakeAccount'),
-        stakeAuthority: resolveAccount('stakeAuthority'),
-        stakeConfigAccount: resolveAccount('stakeConfigAccount'),
-        voteAccount: resolveAccount('voteAccount'),
-    },
-    StakeSplitInstruction: {
-        newSplitAccount: resolveAccount('newSplitAccount'),
-        stakeAccount: resolveAccount('stakeAccount'),
-        stakeAuthority: resolveAccount('stakeAuthority'),
-    },
-    StakeWithdrawInstruction: {
-        destination: resolveAccount('destination'),
-        stakeAccount: resolveAccount('stakeAccount'),
-        withdrawAuthority: resolveAccount('withdrawAuthority'),
-    },
-    StakeDeactivateInstruction: {
-        stakeAccount: resolveAccount('stakeAccount'),
-        stakeAuthority: resolveAccount('stakeAuthority'),
-    },
-    StakeSetLockupInstruction: {
-        custodian: resolveAccount('custodian'),
-        stakeAccount: resolveAccount('stakeAccount'),
-    },
-    StakeMergeInstruction: {
-        destination: resolveAccount('destination'),
-        source: resolveAccount('source'),
-        stakeAuthority: resolveAccount('stakeAuthority'),
-    },
-    StakeAuthorizeWithSeedInstruction: {
-        authorityBase: resolveAccount('authorityBase'),
-        authorityOwner: resolveAccount('authorityOwner'),
-        custodian: resolveAccount('custodian'),
-        newAuthorized: resolveAccount('newAuthorized'),
-        stakeAccount: resolveAccount('stakeAccount'),
-    },
-    StakeInitializeCheckedInstructionDataAuthorized: {
-        staker: resolveAccount('staker'),
-        withdrawer: resolveAccount('withdrawer'),
-    },
-    StakeAuthorizeCheckedInstruction: {
-        authority: resolveAccount('authority'),
-        custodian: resolveAccount('custodian'),
-        newAuthority: resolveAccount('newAuthority'),
-        stakeAccount: resolveAccount('stakeAccount'),
-    },
-    StakeAuthorizeCheckedWithSeedInstruction: {
-        authorityBase: resolveAccount('authorityBase'),
-        authorityOwner: resolveAccount('authorityOwner'),
-        custodian: resolveAccount('custodian'),
-        newAuthorized: resolveAccount('newAuthorized'),
-        stakeAccount: resolveAccount('stakeAccount'),
-    },
-    StakeSetLockupCheckedInstruction: {
-        custodian: resolveAccount('custodian'),
-        stakeAccount: resolveAccount('stakeAccount'),
-    },
-    StakeDeactivateDelinquentInstruction: {
-        referenceVoteAccount: resolveAccount('referenceVoteAccount'),
-        stakeAccount: resolveAccount('stakeAccount'),
-        voteAccount: resolveAccount('voteAccount'),
-    },
-    StakeRedelegateInstruction: {
-        newStakeAccount: resolveAccount('newStakeAccount'),
-        stakeAccount: resolveAccount('stakeAccount'),
-        stakeAuthority: resolveAccount('stakeAuthority'),
-        stakeConfigAccount: resolveAccount('stakeConfigAccount'),
-        voteAccount: resolveAccount('voteAccount'),
-    },
-    CreateAccountInstruction: {
-        newAccount: resolveAccount('newAccount'),
-        owner: resolveAccount('owner'),
-        source: resolveAccount('source'),
-    },
-    AssignInstruction: {
-        account: resolveAccount('account'),
-        owner: resolveAccount('owner'),
-    },
-    TransferInstruction: {
-        destination: resolveAccount('destination'),
-        source: resolveAccount('source'),
-    },
-    CreateAccountWithSeedInstruction: {
-        base: resolveAccount('base'),
-        owner: resolveAccount('owner'),
-        seed: resolveAccount('seed'),
-    },
-    AdvanceNonceAccountInstruction: {
-        nonceAccount: resolveAccount('nonceAccount'),
-        nonceAuthority: resolveAccount('nonceAuthority'),
-    },
-    WithdrawNonceAccountInstruction: {
-        destination: resolveAccount('destination'),
-        nonceAccount: resolveAccount('nonceAccount'),
-        nonceAuthority: resolveAccount('nonceAuthority'),
-    },
-    InitializeNonceAccountInstruction: {
-        nonceAccount: resolveAccount('nonceAccount'),
-        nonceAuthority: resolveAccount('nonceAuthority'),
-    },
-    AuthorizeNonceAccountInstruction: {
-        newAuthorized: resolveAccount('newAuthorized'),
-        nonceAccount: resolveAccount('nonceAccount'),
-        nonceAuthority: resolveAccount('nonceAuthority'),
-    },
-    UpgradeNonceAccountInstruction: {
-        nonceAccount: resolveAccount('nonceAccount'),
-        nonceAuthority: resolveAccount('nonceAuthority'),
-    },
-    AllocateInstruction: {
-        account: resolveAccount('account'),
-    },
-    AllocateWithSeedInstruction: {
-        account: resolveAccount('account'),
-        owner: resolveAccount('owner'),
-    },
-    AssignWithSeedInstruction: {
-        account: resolveAccount('account'),
-        owner: resolveAccount('owner'),
-    },
-    TransferWithSeedInstruction: {
-        destination: resolveAccount('destination'),
-        source: resolveAccount('source'),
-        sourceOwner: resolveAccount('sourceOwner'),
-    },
-    VoteInitializeAccountInstruction: {
-        authorizedVoter: resolveAccount('authorizedVoter'),
-        authorizedWithdrawer: resolveAccount('authorizedWithdrawer'),
-        node: resolveAccount('node'),
-        voteAccount: resolveAccount('voteAccount'),
-    },
-    VoteAuthorizeInstruction: {
-        authority: resolveAccount('authority'),
-        newAuthority: resolveAccount('newAuthority'),
-        voteAccount: resolveAccount('voteAccount'),
-    },
-    VoteAuthorizeWithSeedInstruction: {
-        authorityOwner: resolveAccount('authorityOwner'),
-        newAuthority: resolveAccount('newAuthority'),
-        voteAccount: resolveAccount('voteAccount'),
-    },
-    VoteAuthorizeCheckedWithSeedInstruction: {
-        authorityOwner: resolveAccount('authorityOwner'),
-        newAuthority: resolveAccount('newAuthority'),
-        voteAccount: resolveAccount('voteAccount'),
-    },
-    VoteVoteInstruction: {
-        voteAccount: resolveAccount('voteAccount'),
-        voteAuthority: resolveAccount('voteAuthority'),
-    },
-    VoteUpdateVoteStateInstruction: {
-        voteAccount: resolveAccount('voteAccount'),
-        voteAuthority: resolveAccount('voteAuthority'),
-    },
-    VoteUpdateVoteStateSwitchInstruction: {
-        voteAccount: resolveAccount('voteAccount'),
-        voteAuthority: resolveAccount('voteAuthority'),
-    },
-    VoteCompactUpdateVoteStateInstruction: {
-        voteAccount: resolveAccount('voteAccount'),
-        voteAuthority: resolveAccount('voteAuthority'),
-    },
-    VoteCompactUpdateVoteStateSwitchInstruction: {
-        voteAccount: resolveAccount('voteAccount'),
-        voteAuthority: resolveAccount('voteAuthority'),
-    },
-    VoteWithdrawInstruction: {
-        voteAccount: resolveAccount('voteAccount'),
-        withdrawAuthority: resolveAccount('withdrawAuthority'),
-    },
-    VoteUpdateValidatorIdentityInstruction: {
-        newValidatorIdentity: resolveAccount('newValidatorIdentity'),
-        voteAccount: resolveAccount('voteAccount'),
-        withdrawAuthority: resolveAccount('withdrawAuthority'),
-    },
-    VoteUpdateCommissionInstruction: {
-        voteAccount: resolveAccount('voteAccount'),
-        withdrawAuthority: resolveAccount('withdrawAuthority'),
-    },
-    VoteVoteSwitchInstruction: {
-        voteAccount: resolveAccount('voteAccount'),
-        voteAuthority: resolveAccount('voteAuthority'),
-    },
-    VoteAuthorizeCheckedInstruction: {
-        authority: resolveAccount('authority'),
-        newAuthority: resolveAccount('newAuthority'),
-        voteAccount: resolveAccount('voteAccount'),
-    },
-};
