+++ packages/rpc-graphql/src/schema/transaction.ts
@@ -108,18 +108,3 @@ export const transactionTypeDefs = /* GraphQL */ `
         version: String
     }
 `;
-
-export const transactionResolvers = {
-    Transaction: {
-        __resolveType(transaction: { encoding: string }) {
-            switch (transaction.encoding) {
-                case 'base58':
-                    return 'TransactionBase58';
-                case 'base64':
-                    return 'TransactionBase64';
-                default:
-                    return 'TransactionParsed';
-            }
-        },
-    },
-};
